MAINT: Move PETScKspScheme and LinearTransforScheme to options_parser module
